Adapt to new model
